Moved config stuff outside of utils.py into its own config.py
